Rename the path_stroke module into stroke.
