+++ tessellation/src/stroke.rs
